target/riscv: Add itrigger support when icount is not enabled
When icount is not enabled, there is no API in QEMU that can get the
guest instruction number.

Translate the guest code in a way that each TB only has one instruction.
After the instruction executes, decrese the count by 1 until it reaches 0.
Then the itrigger fires.

Note that only when priviledge matches the itrigger configuration,
the count will decrese.

Signed-off-by: LIU Zhiwei <zhiwei_liu@c-sky.com>